Turn cabin unifier into tkinter program
The cabin unifier has been rewritten as a (GUI-less) tkinter program,
so that users never have to see a terminal window. As the only UI for
the unifier was warnings and prompts, the tkinter version consists of
only message boxes, with a withdrawn root window.

A few missing new line characters were also added to the generated
readme.

+++ library/placeholder files/unifier.py
@@ -1,21 +1,19 @@
+import tkinter as tk
+from tkinter import messagebox
 import configparser, os, sys
+
+root = tk.Tk()
+root.state("withdrawn") # I don't think you need to make a Tk() when just using messageboxes, but if one's created automatically I want to make sure it's hidden
 
 paintjobs_before = 0
 paintjobs_after = 0
 
-if sys.version_info[0] < 3:
-    print("The cabin unifier requires Python 3 to work correctly, you appear to be using Python 2. Please relaunch the unifier using Python 3")
-    input("Unifier cannot continue, press enter to quit")
-    sys.exit()
-
 # read ini file
-uni_ini = configparser.ConfigParser()
-try:
-    uni_ini.read("unifier.ini")
-except FileNotFoundError:
-    print("WARNING - File missing: unifier.ini")
-    input("Unifier cannot continue, press enter to quit")
+if not os.path.isfile("unifier.ini"):
+    messagebox.showerror(title = "File missing", message = "WARNING - File missing: unifier.ini\n\nThe cabin unifier will now quit")
     sys.exit()
+uni_ini = configparser.ConfigParser()
+uni_ini.read("unifier.ini")
 
 internal_name = uni_ini["paintjob"]["internal name"]
 vehicles_to_unify = uni_ini.sections()
@@ -40,16 +38,13 @@ for veh in vehicles:
         try:
             tobj_file = open("vehicle/truck/upgrade/paintjob/{}/{}/cabin_{}.tobj".format(internal_name, veh["make_model"], cab_size), "rb")
         except FileNotFoundError:
-            print("WARNING - File missing: vehicle/truck/upgrade/paintjob/{}/{}/cabin_{}.tobj".format(internal_name, veh["make_model"], cab_size))
-            input("Unifier cannot continue, press enter to quit")
+            messagebox.showerror(title = "File missing", message = "WARNING - File missing: vehicle/truck/upgrade/paintjob/{}/{}/cabin_{}.tobj\n\nThe cabin unifier will now quit".format(internal_name, veh["make_model"], cab_size))
             sys.exit()
         else:
             dds_cab_size = str(tobj_file.read())[-6:-5] # gets just the letter of the DDS referred to, e.g. "b" from cabin_b.dds
             tobj_file.close()
             if dds_cab_size not in veh["cabin_dict"]:
-                print("WARNING - Incorrectly named file: vehicle/truck/upgrade/paintjobs/{}/{}/cabin_{}.dds".format(internal_name, veh["make_model"], dds_cab_size))
-                print("All .dds files need to be named similarly to a .dds file, e.g. \"cabin_b.dds\" is fine, but \"cabin_2.dds\" or \"small_cabin.dds\" is not")
-                print("Unifier cannot continue, press enter to quit")
+                messagebox.showerror(title = "Incorrectly named file", message = "WARNING - Incorrectly named file: vehicle/truck/upgrade/paintjobs/{}/{}/cabin_{}.dds\n\nAll .dds files need to be named similarly to an existing .tobj file, e.g. \"cabin_b.dds\" is fine, but \"cabin_2.dds\" or \"small_cabin.dds\" is not\n\nThe cabin unifier will now quit".format(internal_name, veh["make_model"], dds_cab_size))
                 sys.exit()
             if dds_cab_size not in veh["tobj_cabin_dict"]:
                 veh["tobj_cabin_dict"][dds_cab_size] = []
@@ -58,25 +53,23 @@ for veh in vehicles:
     # check that all the required files exist
     for cab_size in veh["cabin_dict"]:
         if not os.path.isfile("def/vehicle/truck/{}/paint_job/{}_{}.sii".format(veh["path"], internal_name, cab_size)):
-            print("WARNING - File missing: def/vehicle/truck/{}/paint_job/{}_{}.sii".format(veh["path"], internal_name, cab_size))
-            input("Unifier cannot continue, press enter to quit")
+            messagebox.showerror(title = "File missing", message = "WARNING - File missing: def/vehicle/truck/{}/paint_job/{}_{}.sii\n\nThe cabin unifier will now quit".format(veh["path"], internal_name, cab_size))
             sys.exit()
         if veh["accessories"]:
             if not os.path.isfile("def/vehicle/truck/{}/paint_job/accessory/{}_{}.sii".format(veh["path"], internal_name, cab_size)):
-                print("WARNING - File missing: def/vehicle/truck/{}/paint_job/accessory/{}_{}.sii".format(veh["path"], internal_name, cab_size))
-                input("Unifier cannot continue, press enter to quit")
+                messagebox.showerror(title = "File missing", message = "WARNING - File missing: def/vehicle/truck/{}/paint_job/accessory/{}_{}.sii\n\nThe cabin unifier will now quit".format(veh["path"], internal_name, cab_size))
                 sys.exit()
 
 # double check the user wants to proceed before affecting any files, if something goes wrong the blood is on their hands
-print("Looks like everything's in order, ready to unify {} for {} vehicles".format(internal_name, len(vehicles)))
-print("Note that continuing will delete and re-generate a number of .sii files")
-input("Press enter to continue")
+answer = messagebox.askokcancel(title = "Ready to proceed", message = "Looks like everything's in order, ready to unify {} for {} vehicles\n\nNote that continuing will delete and re-generate a number of .sii files\n\nPress OK to continue".format(internal_name, len(vehicles)))
+if not answer:
+    sys.exit()
 
 for veh in vehicles:
     # delete old SIIs
     for cab_size in veh["cabin_dict"]:
         os.remove("def/vehicle/truck/{}/paint_job/{}_{}.sii".format(veh["path"], internal_name, cab_size))
-        if veh["accessories"] and cab_size not in veh["tobj_cab_dict"]: # instead of re-generating the accessory files, simply delete the ones no longer required
+        if veh["accessories"] and cab_size not in veh["tobj_cabin_dict"]: # instead of re-generating the accessory files, simply delete the ones no longer required
             os.remove("def/vehicle/truck/{}/paint_job/accessory/{}_{}.sii".format(veh["path"], internal_name, cab_size))
 
     # generate new SIIs
@@ -95,11 +88,6 @@ for veh in vehicles:
         file.close()
         paintjobs_after += 1
 
-print("")
 percentage_reduction = str(100 * (1 - (paintjobs_after / paintjobs_before))).split(".")[0]
-print("Total paintjobs reduced from {} to {} (reduction of {}%)".format(paintjobs_before, paintjobs_after, percentage_reduction))
-print("")
-print("You may now delete unifier.ini and unifier.py")
-print("If you want to upload your mod to Steam Workshop, you will HAVE to delete them, or the SCS Workshop Uploader will get upset")
-input("Press enter to quit")
+messagebox.showinfo(title = "All done", message = "Total paintjobs reduced from {} to {} (reduction of {}%)\n\nYou may now delete the cabin unifier and unifier.ini\n\nIf you want to upload your mod to Steam Workshop, you will HAVE to delete them, or the SCS Workshop Uploader will get upset\n\nThe cabin unifier will now quit".format(paintjobs_before, paintjobs_after, percentage_reduction))
 sys.exit()
